requirements.txt
pystray
plyer
Pillow